refactor(lifecycle): centralize side effects wiring
