+++ src/lifecycle.ts
@@ -10,7 +10,7 @@ import type {
   RunContext,
   ToolOutputEvent,
 } from "./lifecycle-contract";
-import { POST_EFFECTS, PRE_EFFECTS } from "./lifecycle-effects";
+import { attachLifecycleSideEffects } from "./lifecycle-effects";
 import { phaseFinalize } from "./lifecycle-finalize";
 import { createRunAgent, phaseGenerate } from "./lifecycle-generate";
 import { createLifecyclePolicy } from "./lifecycle-policy";
@@ -21,9 +21,8 @@ import type { MemoryCommitContext, MemoryCommitMetrics } from "./memory-contract
 import { commitDistiller, DISTILLER_PROMPT } from "./memory-distiller";
 import { createInMemoryTaskQueue } from "./task-queue";
 import { renderToolOutputPart } from "./tool-output-content";
-import { DISCOVERY_TOOL_SET, WRITE_TOOL_SET } from "./tool-registry";
+import { WRITE_TOOL_SET } from "./tool-registry";
 import { scopedCallLog } from "./tool-session";
-import { captureUndoBefore, commitUndoCheckpoint, isFilePath, type PendingUndoCapture } from "./undo-checkpoints";
 import { formatWorkspaceCommand, resolveWorkspaceProfile } from "./workspace-profile";
 import { resolveWorkspaceSandboxRoot } from "./workspace-sandbox";
 
@@ -135,82 +134,9 @@ function createRunContext(
   };
 
   session.featureFlags = ctx.features;
-
-  const pendingUndo = new Map<string, PendingUndoCapture>();
-
-  session.onBeforeTool = (preCtx) => runPreEffects(ctx, preCtx);
-  session.onAfterTool = (toolResult) => runPostEffects(ctx, toolResult);
-  session.onBeforeToolAsync = async (preCtx) => {
-    if (!ctx.features.undoCheckpoints) return;
-    if (!WRITE_TOOL_SET.has(preCtx.toolId)) return;
-    const sessionId = ctx.request.sessionId;
-    if (!sessionId || !ctx.workspace) return;
-
-    const args = preCtx.args;
-    const paths: string[] = [];
-    if (preCtx.toolId === "file-edit" || preCtx.toolId === "file-create") {
-      const p = typeof args.path === "string" ? args.path.trim() : "";
-      if (p) paths.push(p);
-    } else if (preCtx.toolId === "file-delete") {
-      const ps = args.paths;
-      if (Array.isArray(ps)) {
-        for (const p of ps) if (typeof p === "string" && p.trim().length > 0) paths.push(p.trim());
-      }
-    } else if (preCtx.toolId === "code-edit") {
-      const p = typeof args.path === "string" ? args.path.trim() : "";
-      if (p && (await isFilePath(ctx.workspace, p))) paths.push(p);
-    }
-    if (paths.length === 0) return;
-
-    const capture = await captureUndoBefore({
-      workspace: ctx.workspace,
-      toolCallId: preCtx.toolCallId,
-      toolId: preCtx.toolId,
-      paths,
-    });
-    pendingUndo.set(preCtx.toolCallId, capture);
-  };
-
-  session.onAfterToolAsync = async (postCtx) => {
-    if (!ctx.features.undoCheckpoints) return;
-    if (!WRITE_TOOL_SET.has(postCtx.toolId)) return;
-    const sessionId = ctx.request.sessionId;
-    if (!sessionId || !ctx.workspace) return;
-
-    const pending = pendingUndo.get(postCtx.toolCallId);
-    if (!pending) return;
-    pendingUndo.delete(postCtx.toolCallId);
-    await commitUndoCheckpoint({ workspace: ctx.workspace, sessionId, pending });
-  };
+  attachLifecycleSideEffects(ctx, session);
 
   return ctx;
-}
-
-function runPreEffects(
-  ctx: RunContext,
-  { toolId }: { toolId: string; args: Record<string, unknown> },
-): { append: string } | undefined {
-  if (DISCOVERY_TOOL_SET.has(toolId)) return undefined;
-  for (const effect of PRE_EFFECTS) {
-    effect.run(ctx);
-  }
-  return undefined;
-}
-
-function runPostEffects(
-  ctx: RunContext,
-  { toolId, args }: { toolId: string; args: Record<string, unknown> },
-): { append: string } | undefined {
-  if (!WRITE_TOOL_SET.has(toolId)) return undefined;
-  const path = typeof args.path === "string" ? args.path.trim() : "";
-  if (!path) return undefined;
-  const paths = [path];
-  let lintOutput: string | undefined;
-  for (const effect of POST_EFFECTS) {
-    const result = effect.run(ctx, paths);
-    if (result.lintOutput) lintOutput = result.lintOutput;
-  }
-  return lintOutput ? { append: `Lint errors:\n${lintOutput}` } : undefined;
 }
 
 export function resolveSignal(ctx: RunContext): LifecycleSignal | undefined {
